# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Activity {
  id: number;
  title: string;
  detail: string;
  time: string;
  tone: string;
}

export interface Dashboard {
  studentCount: number;
  teacherCount: number;
  attendance: number;
  upcomingExams: number;
  recentActivity: Activity[];
}

export interface Student {
  id: number;
  name: string;
  admissionNumber: string;
  className: string;
  section: string;
  rollNumber: number;
  attendance: number;
  averageScore: number;
  achievementCount: number;
  /** @nullable */
  parentContact: string | null;
  /** @nullable */
  studentContact: string | null;
  /** @nullable */
  address: string | null;
  /** @nullable */
  email: string | null;
  avatar?: string;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  admissionNumber: string;
  className: string;
  section: string;
  rollNumber: number;
  parentContact: string;
  studentContact: string;
  address: string;
  email: string;
}

export interface Profile {
  id: string;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  studentId: number | null;
}

export interface Teacher {
  id: number;
  name: string;
  subject: string;
  specialization: string;
  email: string;
  phone: string;
  experience: string;
  initials: string;
  color: string;
}

export interface Resource {
  id: number;
  title: string;
  kind: string;
  subject: string;
  className: string;
  year: number;
  pages: number;
  downloads: number;
}

export interface Notice {
  id: number;
  title: string;
  body: string;
  category: string;
  publishedAt: string;
  author: string;
  pinned: boolean;
}

export interface NoticeInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  category: string;
  pinned?: boolean;
}

export interface SchoolEvent {
  id: number;
  title: string;
  date: string;
  type: string;
  description: string;
  color: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  date: string;
  type: string;
  description: string;
  color?: string;
}

export interface Lecture {
  id: number;
  title: string;
  teacher: string;
  subject: string;
  className: string;
  duration: string;
  scheduledAt: string;
  status: string;
  meetingUrl: string;
}

export type ListStudentsParams = {
search?: string;
className?: string;
};

export type ListResourcesParams = {
kind?: string;
className?: string;
};