# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  admissionNumber: string;
  className: string;
  section: string;
  rollNumber: number;
  attendance: number;
  averageScore: number;
  achievementCount: number;
  /** @nullable */
  parentContact: string | null;
  /** @nullable */
  studentContact: string | null;
  /** @nullable */
  address: string | null;
  /** @nullable */
  email: string | null;
  avatar?: string;
}